+++ views_pipeline_core/data/utils.py
@@ -2,6 +2,8 @@ import logging
 import numpy as np
 import requests
 import json
+import pandas as pd
+from typing import Union
 
 logger = logging.getLogger(__name__)
 
@@ -68,3 +70,54 @@ def convert_json_to_list_of_dicts(json_data):
     # Convert each JSON object to a dictionary and store in a list
     list_of_dicts = [json.loads(obj) for obj in json_objects]
     return list_of_dicts
+
+def replace_nan_values(df: pd.DataFrame, value: Union[int, float] = 0.0) -> pd.DataFrame:
+        """
+        Replaces NaN values in a DataFrame or a list of DataFrames with a specified value.
+
+        This function iterates through the columns of the input DataFrame(s) and replaces
+        NaN values with the provided value. For object-type columns containing lists, it
+        replaces NaN values within the lists.
+
+        Args:
+            df (Union[pd.DataFrame, List[pd.DataFrame]]): The input DataFrame or a list of DataFrames
+                where NaN values need to be replaced.
+            value (Union[int, float], optional): The value to replace NaN values with. Defaults to 0.0.
+
+        Returns:
+            Union[pd.DataFrame, List[pd.DataFrame]]: The DataFrame or list of DataFrames with NaN
+            values replaced.
+
+        Raises:
+            ValueError: If the input is neither a DataFrame nor a list of DataFrames.
+
+        Examples:
+            >>> import pandas as pd
+            >>> import numpy as np
+            >>> df = pd.DataFrame({'A': [1, np.nan, 3], 'B': [np.nan, 2, 3]})
+            >>> replace_nan_values(df, value=0)
+               A    B
+            0  1.0  0.0
+            1  0.0  2.0
+            2  3.0  3.0
+        """
+        def _replace_nan(d):
+            for column in d.columns:
+                if d[column].dtype == "object":
+                    d[column] = d[column].apply(
+                        lambda x: [np.float64(value) if np.isnan(i) else float(i) for i in x]
+                    )
+                else:
+                    d[column] = d[column].apply(
+                        lambda x: np.float64(value) if np.isnan(x) else float(x)
+                    )
+            return d
+        if isinstance(df, list):
+            for i, d in enumerate(df):
+                df[i] = _replace_nan(d)
+            return df
+        elif isinstance(df, pd.DataFrame):
+            df = _replace_nan(df)
+            return df
+        else:
+            raise ValueError(f"Input should be a DataFrame or a list of DataFrames. Found {type(df)}")
